feat: refactor status bar
